requirements.txt
fastapi
uvicorn
openai
openai-agents
websockets
python-dotenv
